fix(evm): refactor evm precompiled contract into a separate library, refactor of sdk and entrypoint (panic handler and allocator) that is easier to use now
