+++ crates/evm/src/evm.rs
@@ -1,5 +1,11 @@
-use crate::evm::EVM;
-use fluentbase_sdk::SharedAPI;
+use crate::{
+    bytecode::AnalyzedBytecode,
+    gas::Gas,
+    memory::SharedMemory,
+    result::{InstructionResult, InterpreterResult},
+    stack::Stack,
+};
+use fluentbase_sdk::{Bytes, ContractContextReader, SharedAPI};
 
 mod arithmetic;
 mod bitwise;
@@ -7,6 +13,7 @@ mod context;
 mod contract;
 mod control;
 mod host;
+mod i256;
 mod memory;
 mod stack;
 mod system;
@@ -16,6 +23,66 @@ pub type Instruction<SDK> = fn(&mut EVM<SDK>);
 
 /// Instruction table is a list of instruction function pointers mapped to 256 EVM opcodes.
 pub type InstructionTable<SDK> = [Instruction<SDK>; 0x100];
+
+pub struct EVM<'a, SDK: SharedAPI> {
+    pub(crate) sdk: &'a mut SDK,
+    pub(crate) analyzed_bytecode: AnalyzedBytecode,
+    pub(crate) input: &'a [u8],
+    pub(crate) gas: Gas,
+    pub(crate) ip: *const u8,
+    pub(crate) state: InstructionResult,
+    pub(crate) return_data_buffer: Bytes,
+    pub(crate) is_static: bool,
+    pub(crate) output: Option<InterpreterResult>,
+    pub(crate) memory: SharedMemory,
+    pub(crate) stack: Stack,
+}
+
+impl<'a, SDK: SharedAPI> EVM<'a, SDK> {
+    pub fn new(sdk: &'a mut SDK, bytecode: &'a [u8], input: &'a [u8], gas_limit: u64) -> Self {
+        let is_static = sdk.context().contract_is_static();
+        let analyzed_bytecode = AnalyzedBytecode::new(bytecode);
+        let ip = analyzed_bytecode.bytecode.as_ptr();
+        let gas = Gas::new(gas_limit);
+        Self {
+            sdk,
+            analyzed_bytecode,
+            input,
+            gas,
+            ip,
+            state: InstructionResult::Continue,
+            return_data_buffer: Default::default(),
+            is_static,
+            output: None,
+            memory: Default::default(),
+            stack: Default::default(),
+        }
+    }
+
+    pub fn exec(&mut self) -> InterpreterResult {
+        let instruction_table = make_instruction_table::<SDK>();
+        while self.state == InstructionResult::Continue {
+            let opcode = unsafe { *self.ip };
+            self.ip = unsafe { self.ip.offset(1) };
+            instruction_table[opcode as usize](self);
+        }
+        if let Some(output) = self.output.take() {
+            return output;
+        }
+        InterpreterResult {
+            result: self.state,
+            output: Bytes::new(),
+            gas: self.gas,
+        }
+    }
+
+    pub fn program_counter(&self) -> usize {
+        unsafe {
+            self.ip
+                .offset_from(self.analyzed_bytecode.bytecode.as_ptr()) as usize
+        }
+    }
+}
 
 #[inline]
 pub const fn make_instruction_table<SDK: SharedAPI>() -> InstructionTable<SDK> {
