+++ crates/sdk/src/allocator.rs
@@ -1,9 +1,8 @@
 use alloc::vec::Vec;
-#[cfg(target_arch = "wasm32")]
-use core::alloc::GlobalAlloc;
 
 const WASM_PAGE_SIZE_IN_BYTES: usize = 65536;
 
+#[allow(dead_code)]
 fn calc_pages_needed(pages_allocated: usize, ptr: usize) -> usize {
     let current_memory = pages_allocated * WASM_PAGE_SIZE_IN_BYTES;
     if ptr >= current_memory {
@@ -23,36 +22,6 @@ fn test_pages_needed() {
     assert_eq!(calc_pages_needed(1, 65535), 0);
     assert_eq!(calc_pages_needed(1, 65536 * 2), 3);
     assert_eq!(calc_pages_needed(5, 327680), 6);
-}
-
-#[cfg(target_arch = "wasm32")]
-fn _sys_alloc_aligned(bytes: usize, align: usize) -> *mut u8 {
-    extern "C" {
-        static __heap_base: u8;
-    }
-    static mut HEAP_POS: usize = 0;
-    let mut heap_pos = unsafe { HEAP_POS };
-    if heap_pos == 0 {
-        heap_pos = unsafe { (&__heap_base) as *const u8 as usize };
-    }
-    let offset = heap_pos & (align - 1);
-    if offset != 0 {
-        heap_pos += align - offset;
-    }
-    // allocate memory pages if needed
-    let pages_allocated = core::arch::wasm32::memory_size::<0>();
-    let pages_needed = calc_pages_needed(pages_allocated, heap_pos + bytes);
-    if pages_needed > 0 {
-        let new_pages = core::arch::wasm32::memory_grow::<0>(pages_needed);
-        if new_pages == usize::MAX {
-            unreachable!("out of memory");
-        }
-    }
-    // return allocated pointer
-    let ptr = heap_pos as *mut u8;
-    heap_pos += bytes;
-    unsafe { HEAP_POS = heap_pos };
-    ptr
 }
 
 #[inline(always)]
@@ -75,10 +44,37 @@ pub fn alloc_vec(len: usize) -> Vec<u8> {
 pub struct HeapBaseAllocator {}
 
 #[cfg(target_arch = "wasm32")]
-unsafe impl GlobalAlloc for HeapBaseAllocator {
+unsafe impl core::alloc::GlobalAlloc for HeapBaseAllocator {
     #[inline(always)]
     unsafe fn alloc(&self, layout: core::alloc::Layout) -> *mut u8 {
-        _sys_alloc_aligned(layout.size(), layout.align())
+        let bytes: usize = layout.size();
+        let align: usize = layout.align();
+        extern "C" {
+            static __heap_base: u8;
+        }
+        static mut HEAP_POS: usize = 0;
+        let mut heap_pos = unsafe { HEAP_POS };
+        if heap_pos == 0 {
+            heap_pos = unsafe { (&__heap_base) as *const u8 as usize };
+        }
+        let offset = heap_pos & (align - 1);
+        if offset != 0 {
+            heap_pos += align - offset;
+        }
+        // allocate memory pages if needed
+        let pages_allocated = core::arch::wasm32::memory_size::<0>();
+        let pages_needed = calc_pages_needed(pages_allocated, heap_pos + bytes);
+        if pages_needed > 0 {
+            let new_pages = core::arch::wasm32::memory_grow::<0>(pages_needed);
+            if new_pages == usize::MAX {
+                unreachable!("out of memory");
+            }
+        }
+        // return allocated pointer
+        let ptr = heap_pos as *mut u8;
+        heap_pos += bytes;
+        unsafe { HEAP_POS = heap_pos };
+        ptr
     }
 
     #[inline(always)]
