+++ examples/simple-storage/Cargo.toml
@@ -1,7 +1,13 @@
 [package]
 name = "fluentbase-examples-simple-storage"
 version = "0.1.0"
-edition = "2021"
+authors.workspace = true
+repository.workspace = true
+edition.workspace = true
+readme.workspace = true
+license.workspace = true
+keywords.workspace = true
+categories.workspace = true
 
 [dependencies]
 fluentbase-sdk = { workspace = true }
